just keep typing

+++ java/tenmo-server/src/main/java/com/techelevator/tenmo/model/Transfers.java
@@ -1,16 +1,18 @@
 package com.techelevator.tenmo.model;
 
+import java.math.BigDecimal;
+
 public class Transfers {
 	private int transfer_id;
 	private int transfer_type_id;
 	private int transfer_status_id;
 	private int account_from;
 	private int account_to;
-	private double amount;
+	private BigDecimal amount;
 	
 	public Transfers() {}
 	
-	public Transfers(int transfer_id, int transfer_type_id, int transfer_status_id, int account_from, int account_to, double amount) {
+	public Transfers(int transfer_id, int transfer_type_id, int transfer_status_id, int account_from, int account_to, BigDecimal amount) {
 		this.transfer_id = transfer_id;
 		this.transfer_type_id = transfer_type_id;
 		this.transfer_status_id = transfer_status_id;
@@ -59,11 +61,11 @@ public class Transfers {
 		this.account_to = account_to;
 	}
 
-	public double getAmount() {
+	public BigDecimal getAmount() {
 		return amount;
 	}
 
-	public void setAmount(double amount) {
+	public void setAmount(BigDecimal amount) {
 		this.amount = amount;
 	}
